Different method for setting password of hacluster
It seems that in openSUSE 12.3, passwd no longer accepts --stdin
as a command line parameter. This should work even without the
command line flag support.

One problem is that passwd writes quite a bit of output to stderr,
so I'm redirecting both stdin and stderr to /dev/null. If there is
a problem with setting the password, it will be more difficult to
figure out.