Sync all config files across all nodes when adding new node (bnc#867048)
Previously we used `csync2 -xv -P $(hostname)` to sync files only to the
new node.  Unfortunately this leaves all other files marked dirty on
that node, so a subsequent join of another node may have a failed csync2
run, causing e.g.: corosync.conf to be out of date at that point on one
node.  To fix this, we just sync all files across the cluster when doing
a join.  *sigh*

Signed-off-by: Tim Serong <tserong@suse.com>