Use fdisk to limit partprobe to valid devices (bnc#885287)
This function uses fdisk to create a list of valid devices for
probing with partprobe.  This prevents partprobe from failing on
read-only mounted devices such as /dev/sr0 (etc) that might cause
it to return an error when it exits.  This allows partprobe to
run without forcing _die to bail out.